phraseRendering.tsの処理をsinging.tsに移動（移動のみ）
